[mlir][gpu] Separate the barrier elimination code from transform ops
Allows the barrier elimination code to be run from C++ as well. The code
from transforms dialect is copied as-is, the pass and populate functions
have beed added at the end.
