+++ mlir/include/mlir/Dialect/GPU/Transforms/Passes.h
@@ -165,6 +165,9 @@ void populateGpuDecomposeMemrefsPatterns(RewritePatternSet &patterns);
 /// Pass decomposes memref ops inside `gpu.launch` body.
 std::unique_ptr<Pass> createGpuDecomposeMemrefsPass();
 
+/// Erase barriers that do not enforce conflicting memory side effects.
+void populateGpuEliminateBarriersPatterns(RewritePatternSet &patterns);
+
 /// Generate the code for registering passes.
 #define GEN_PASS_REGISTRATION
 #include "mlir/Dialect/GPU/Transforms/Passes.h.inc"
